requirements.txt
pbr>=0.5.21,<1.0

urwid
docutils>=0.12
pygments
pillow # or PIL.  Optional; only used for images.

